+++ src/Dap.Remote.Squidex.Sync/_Gen/Types.fs
@@ -14,6 +14,7 @@ open Dap.Remote.Squidex
 type SyncSnapshot = {
     Id : (* SyncSnapshot *) string
     Time : (* SyncSnapshot *) Instant
+    Queries : (* SyncSnapshot *) Map<string, string>
     Contents : (* SyncSnapshot *) Map<string, ContentsWithTotalResult>
     Errors : (* SyncSnapshot *) Map<string, string>
 } with
@@ -21,6 +22,7 @@ type SyncSnapshot = {
         (
             ?id : (* SyncSnapshot *) string,
             ?time : (* SyncSnapshot *) Instant,
+            ?queries : (* SyncSnapshot *) Map<string, string>,
             ?contents : (* SyncSnapshot *) Map<string, ContentsWithTotalResult>,
             ?errors : (* SyncSnapshot *) Map<string, string>
         ) : SyncSnapshot =
@@ -29,6 +31,8 @@ type SyncSnapshot = {
                 |> Option.defaultWith (fun () -> "")
             Time = (* SyncSnapshot *) time
                 |> Option.defaultWith (fun () -> (getNow' ()))
+            Queries = (* SyncSnapshot *) queries
+                |> Option.defaultWith (fun () -> Map.empty)
             Contents = (* SyncSnapshot *) contents
                 |> Option.defaultWith (fun () -> Map.empty)
             Errors = (* SyncSnapshot *) errors
@@ -38,6 +42,8 @@ type SyncSnapshot = {
         {this with Id = id}
     static member SetTime ((* SyncSnapshot *) time : Instant) (this : SyncSnapshot) =
         {this with Time = time}
+    static member SetQueries ((* SyncSnapshot *) queries : Map<string, string>) (this : SyncSnapshot) =
+        {this with Queries = queries}
     static member SetContents ((* SyncSnapshot *) contents : Map<string, ContentsWithTotalResult>) (this : SyncSnapshot) =
         {this with Contents = contents}
     static member SetErrors ((* SyncSnapshot *) errors : Map<string, string>) (this : SyncSnapshot) =
@@ -47,6 +53,7 @@ type SyncSnapshot = {
             E.object [
                 "id", E.string (* SyncSnapshot *) this.Id
                 "time", E.instant (* SyncSnapshot *) this.Time
+                "queries", (E.dict E.string) (* SyncSnapshot *) this.Queries
                 "contents", (E.dict ContentsWithTotalResult.JsonEncoder) (* SyncSnapshot *) this.Contents
                 "errors", (E.dict E.string) (* SyncSnapshot *) this.Errors
             ]
@@ -55,6 +62,7 @@ type SyncSnapshot = {
             {
                 Id = get.Required.Field (* SyncSnapshot *) "id" D.string
                 Time = get.Required.Field (* SyncSnapshot *) "time" D.instant
+                Queries = get.Required.Field (* SyncSnapshot *) "queries" (D.dict D.string)
                 Contents = get.Required.Field (* SyncSnapshot *) "contents" (D.dict ContentsWithTotalResult.JsonDecoder)
                 Errors = get.Required.Field (* SyncSnapshot *) "errors" (D.dict D.string)
             }
@@ -68,10 +76,40 @@ type SyncSnapshot = {
         this |> SyncSnapshot.SetId id
     member this.WithTime ((* SyncSnapshot *) time : Instant) =
         this |> SyncSnapshot.SetTime time
+    member this.WithQueries ((* SyncSnapshot *) queries : Map<string, string>) =
+        this |> SyncSnapshot.SetQueries queries
     member this.WithContents ((* SyncSnapshot *) contents : Map<string, ContentsWithTotalResult>) =
         this |> SyncSnapshot.SetContents contents
     member this.WithErrors ((* SyncSnapshot *) errors : Map<string, string>) =
         this |> SyncSnapshot.SetErrors errors
+
+(*
+ * Generated: <Union>
+ *     IsJson
+ *)
+type SyncResult =
+    | Succeed of snapshot : SyncSnapshot
+    | Failed of error : string
+with
+    static member CreateSucceed snapshot : SyncResult =
+        Succeed (snapshot)
+    static member CreateFailed error : SyncResult =
+        Failed (error)
+    static member JsonSpec' : CaseSpec<SyncResult> list =
+        [
+            CaseSpec<SyncResult>.Create ("Succeed", [
+                SyncSnapshot.JsonSpec
+            ])
+            CaseSpec<SyncResult>.Create ("Failed", [
+                S.string
+            ])
+        ]
+    static member JsonEncoder = E.union SyncResult.JsonSpec'
+    static member JsonDecoder = D.union SyncResult.JsonSpec'
+    static member JsonSpec =
+        FieldSpec.Create<SyncResult> (SyncResult.JsonEncoder, SyncResult.JsonDecoder)
+    interface IJson with
+        member this.ToJson () = SyncResult.JsonEncoder this
 
 (*
  * Generated: <Combo>
@@ -83,6 +121,7 @@ type SyncProps (owner : IOwner, key : Key) =
     let reloadInterval = target'.AddVar<(* SyncProps *) Duration option> ((E.option E.duration), (D.option D.duration), "reload_interval", None, None)
     let snapshots = target'.AddDict<(* SyncProps *) SyncSnapshot> (SyncSnapshot.JsonEncoder, SyncSnapshot.JsonDecoder, "snapshots", (SyncSnapshot.Create ()), None)
     let loading = target'.AddVar<(* SyncProps *) bool> (E.bool, D.bool, "loading", false, None)
+    let lastSnapshotId = target'.AddVar<(* SyncProps *) string> (E.string, D.string, "last_snapshot_id", "", None)
     let lastLoadedTime = target'.AddVar<(* SyncProps *) Instant> (E.instant, D.instant, "last_loaded_time", (getNow' ()), None)
     do (
         base.Setup (target')
@@ -98,6 +137,7 @@ type SyncProps (owner : IOwner, key : Key) =
     member __.ReloadInterval (* SyncProps *) : IVarProperty<Duration option> = reloadInterval
     member __.Snapshots (* SyncProps *) : IDictProperty<IVarProperty<SyncSnapshot>> = snapshots
     member __.Loading (* SyncProps *) : IVarProperty<bool> = loading
+    member __.LastSnapshotId (* SyncProps *) : IVarProperty<string> = lastSnapshotId
     member __.LastLoadedTime (* SyncProps *) : IVarProperty<Instant> = lastLoadedTime
 
 (*
@@ -109,7 +149,7 @@ type ISyncConfig =
     abstract SyncProps : SyncProps with get
     abstract OnLoaded : IChannel<SyncSnapshot> with get
     abstract GetNextSnapshotId : IHandler<unit, string> with get
-    abstract ReloadAsync : IAsyncHandler<unit, bool> with get
+    abstract ReloadAsync : IAsyncHandler<unit, SyncResult> with get
 
 (*
  * Generated: <Context>
@@ -122,15 +162,15 @@ type BaseSyncConfig<'context when 'context :> ISyncConfig> (logging : ILogging) 
     inherit CustomContext<'context, ContextSpec<SyncProps>, SyncProps> (logging, new ContextSpec<SyncProps>(SyncConfigKind, SyncProps.Create))
     let onLoaded = base.Channels.Add<SyncSnapshot> (SyncSnapshot.JsonEncoder, SyncSnapshot.JsonDecoder, "on_loaded")
     let getNextSnapshotId = base.Handlers.Add<unit, string> (E.unit, D.unit, E.string, D.string, "get_next_snapshot_id")
-    let reloadAsync = base.AsyncHandlers.Add<unit, bool> (E.unit, D.unit, E.bool, D.bool, "reload")
+    let reloadAsync = base.AsyncHandlers.Add<unit, SyncResult> (E.unit, D.unit, SyncResult.JsonEncoder, SyncResult.JsonDecoder, "reload")
     member this.SyncProps : SyncProps = this.Properties
     member __.OnLoaded : IChannel<SyncSnapshot> = onLoaded
     member __.GetNextSnapshotId : IHandler<unit, string> = getNextSnapshotId
-    member __.ReloadAsync : IAsyncHandler<unit, bool> = reloadAsync
+    member __.ReloadAsync : IAsyncHandler<unit, SyncResult> = reloadAsync
     interface ISyncConfig with
         member this.SyncProps : SyncProps = this.Properties
         member __.OnLoaded : IChannel<SyncSnapshot> = onLoaded
         member __.GetNextSnapshotId : IHandler<unit, string> = getNextSnapshotId
-        member __.ReloadAsync : IAsyncHandler<unit, bool> = reloadAsync
+        member __.ReloadAsync : IAsyncHandler<unit, SyncResult> = reloadAsync
     interface IFeature
     member this.AsSyncConfig = this :> ISyncConfig