+++ src/Dap.Remote.Squidex.Sync/RELEASE_NOTES.md
@@ -1,5 +1,14 @@
+## 0.2.4
+* Add SyncResult
+
+## 0.2.3
+* Add queries to SyncSnapshot
+
+## 0.2.2
+* Add SyncSnapshot.CastContent<> method
+
 ## 0.2.1
-- Follow Dap.Remote.Squidex changes
+* Follow Dap.Remote.Squidex changes
 
 ## 0.1.0
-- Basic logic to sync data from squidex server
+* Basic logic to sync data from squidex server